+++ src/array/linestring/mod.rs
@@ -2,9 +2,11 @@
 //! geometries.
 
 pub use array::LineStringArray;
-pub use builder::{LineStringBuilder, LineStringCapacity};
+pub use builder::LineStringBuilder;
+pub use capacity::LineStringCapacity;
 pub use iterator::LineStringArrayIter;
 
 mod array;
 pub(crate) mod builder;
+mod capacity;
 pub mod iterator;
