+++ src/array/mixed/mod.rs
@@ -1,7 +1,9 @@
 pub use array::MixedGeometryArray;
-pub use builder::{MixedCapacity, MixedGeometryBuilder};
+pub use builder::MixedGeometryBuilder;
+pub use capacity::MixedCapacity;
 pub use iterator::MixedGeometryArrayIter;
 
 pub mod array;
 pub mod builder;
+mod capacity;
 mod iterator;
