+++ src/array/multilinestring/mod.rs
@@ -2,9 +2,11 @@
 //! MultiLineString geometries.
 
 pub use array::MultiLineStringArray;
-pub use builder::{MultiLineStringBuilder, MultiLineStringCapacity};
+pub use builder::MultiLineStringBuilder;
+pub use capacity::MultiLineStringCapacity;
 pub use iterator::MultiLineStringArrayIter;
 
 mod array;
 mod builder;
+mod capacity;
 pub mod iterator;
