+++ src/array/multipolygon/mod.rs
@@ -2,9 +2,11 @@
 //! geometries.
 
 pub use array::MultiPolygonArray;
-pub use builder::{MultiPolygonBuilder, MultiPolygonCapacity};
+pub use builder::MultiPolygonBuilder;
+pub use capacity::MultiPolygonCapacity;
 pub use iterator::MultiPolygonArrayIter;
 
 mod array;
 mod builder;
+mod capacity;
 pub mod iterator;
